requirements.txt
fastapi
pydantic
python-dotenv
langchain_community
langchain_openai
langchain_chroma
langchain
uvicorn 
pypdf
chromadb
# postgre
sqlalchemy 
psycopg2-binary